requirements.txt
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
openai>=1.30.0
transformers>=4.40.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
